Change name to python_ghost_cursor
