+++ python_ghost_cursor/js/mouseHelper.js
