+++ python_ghost_cursor/playwright/async_api/_spoof.py
@@ -1,16 +1,15 @@
 import asyncio
 import logging
-import time
 import random
-from typing import Union, Optional, Dict, List
-from playwright.sync_api import Page, ElementHandle
+from typing import Union, Coroutine, Optional, Dict, List
+from playwright.async_api import Page, ElementHandle, CDPSession
 
-from pyppeteer_ghost_cursor.shared._math import (
+from python_ghost_cursor.shared._math import (
     Vector,
     origin,
     overshoot,
 )
-from pyppeteer_ghost_cursor.shared._spoof import (
+from python_ghost_cursor.shared._spoof import (
     path,
     should_overshoot,
     get_random_box_point,
@@ -27,12 +26,16 @@ class GhostCursor:
         self.moving = False
         self.overshoot_spread = 10
         self.overshoot_radius = 120
-        self.cdp_session = page.context.new_cdp_session(page)
 
-    def get_random_page_point(self) -> Vector:
+    async def get_cdp_session(self) -> Coroutine[None, None, CDPSession]:
+        if not hasattr(self, "cdp_session"):
+            self.cdp_session = await self.page.context.new_cdp_session(self.page)
+        return self.cdp_session
+
+    async def get_random_page_point(self) -> Coroutine[None, None, Vector]:
         """Get a random point on a browser window"""
         target_id = self.page.target._targetId
-        window = self.cdp_session.send(
+        window = (await self.get_cdp_session()).send(
             "Browser.getWindowForTarget", {"targetId": target_id}
         )
         return get_random_box_point(
@@ -48,8 +51,8 @@ class GhostCursor:
         """Start random mouse movements. Function recursively calls itself"""
         try:
             if not self.moving:
-                rand = self.get_random_page_point()
-                self.trace_path(path(self.previous, rand), True)
+                rand = await self.get_random_page_point()
+                await self.trace_path(path(self.previous, rand), True)
                 self.previous = rand
             await asyncio.sleep(random.random() * 2)
             asyncio.ensure_future(
@@ -58,59 +61,59 @@ class GhostCursor:
         except:
             logger.debug("Warning: stopping random mouse movements")
 
-    def trace_path(self, vectors: List[Vector], abort_on_move: bool = False) -> None:
+    async def trace_path(self, vectors: List[Vector], abort_on_move: bool = False):
         """Move the mouse over a number of vectors"""
         for v in vectors:
             try:
                 # In case this is called from random mouse movements and the users wants to move the mouse, abort
                 if abort_on_move and self.moving:
                     return
-                self.page.mouse.move(v.x, v.y)
+                await self.page.mouse.move(v.x, v.y)
                 self.previous = v
             except Exception as exc:
                 # Exit function if the browser is no longer connected
-                if not self.page.browser.is_connected():
+                if not (await self.page.browser.is_connected()):
                     return
                 logger.debug("Warning: could not move mouse, error message: %s", exc)
 
     def toggle_random_move(self, random_: bool):
         self.moving = not random_
 
-    def click(
+    async def click(
         self,
         selector: Optional[Union[str, ElementHandle]],
         padding_percentage: Optional[float] = None,
         wait_for_selector: Optional[float] = None,
         wait_for_click: Optional[float] = None,
-    ) -> None:
+    ):
         self.toggle_random_move(False)
         if selector is not None:
-            self.move(selector, padding_percentage, wait_for_selector)
+            await self.move(selector, padding_percentage, wait_for_selector)
             self.toggle_random_move(False)
 
         try:
-            self.page.mouse.down()
+            await self.page.mouse.down()
             if wait_for_click is not None:
-                time.sleep(wait_for_click / 1000)
-            self.page.mouse.up()
+                await asyncio.sleep(wait_for_click / 1000)
+            await self.page.mouse.up()
         except Exception as exc:
             logger.debug("Warning: could not click mouse, error message: %s", exc)
 
-        time.sleep(random.random() * 2)
+        await asyncio.sleep(random.random() * 2)
         self.toggle_random_move(True)
 
-    def move(
+    async def move(
         self,
         selector: Union[str, ElementHandle],
         padding_percentage: Optional[float] = None,
         wait_for_selector: Optional[float] = None,
-    ) -> None:
+    ):
         self.toggle_random_move(False)
         elem = None
         if isinstance(selector, str):
             if wait_for_selector:
-                self.page.wait_for_selector(selector, timeout=wait_for_selector)
-            elem = self.page.query_selector(selector)
+                await self.page.wait_for_selector(selector, timeout=wait_for_selector)
+            elem = await self.page.query_selector(selector)
             if elem is None:
                 raise Exception(
                     'Could not find element with selector "${}", make sure you\'re waiting for the elements with "puppeteer.wait_for_selector"'.format(
@@ -121,8 +124,8 @@ class GhostCursor:
             elem = selector
 
         # Make sure the object is in view
-        elem.scroll_into_view_if_needed()
-        box = elem.bounding_box()
+        await elem.scroll_into_view_if_needed()
+        box = await elem.bounding_box()
         if box is None:
             raise Exception(
                 "Could not find the dimensions of the element you're clicking on, this might be a bug?"
@@ -135,7 +138,7 @@ class GhostCursor:
             if overshooting
             else destination
         )
-        self.trace_path(path(self.previous, to))
+        await self.trace_path(path(self.previous, to))
 
         if overshooting:
             bounding_box = {
@@ -145,14 +148,14 @@ class GhostCursor:
                 "y": destination.y,
             }
             correction = path(to, bounding_box, self.overshoot_spread)
-            self.trace_path(correction)
+            await self.trace_path(correction)
         self.previous = destination
         self.toggle_random_move(True)
 
-    def move_to(self, destination: dict) -> None:
+    async def move_to(self, destination: dict):
         destination_vector = Vector(destination["x"], destination["y"])
         self.toggle_random_move(False)
-        self.trace_path(path(self.previous, destination_vector))
+        await self.trace_path(path(self.previous, destination_vector))
         self.toggle_random_move(True)
