+++ python_ghost_cursor/pyppeteer/_spoof.py
@@ -13,12 +13,12 @@ except ModuleNotFoundError:
     # https://github.com/pyppeteer/pyppeteer/blob/dev/pyppeteer/element_handle.py
     from pyppeteer.element_handle import ElementHandle
 
-from pyppeteer_ghost_cursor.shared._math import (
+from python_ghost_cursor.shared._math import (
     Vector,
     origin,
     overshoot,
 )
-from pyppeteer_ghost_cursor.shared._spoof import (
+from python_ghost_cursor.shared._spoof import (
     path,
     should_overshoot,
     get_random_box_point,
